pyproject.toml
[project]
name = "pyakmotorctrl"
version = "0.1.0"
description = "Python library to control AK60-6 motor via CAN in servo mode"
readme = "README.md"
license = {text = "LGPL-3.0-or-later"}
requires-python = ">=3.12"
dependencies = [
    "python-can>=4.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
    "ty>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
